+++ DZObjectBuilder/io/__init__.py
@@ -13,8 +13,6 @@ if "binary_handler" in locals():
         reload(data_p3d)
     if "data_p3d_odol" in locals():
         reload(data_p3d_odol)
-    if "data_rtm" in locals():
-        reload(data_rtm)
     if "data_tbcsv" in locals():
         reload(data_tbcsv)
     if "data_paa" in locals():
@@ -25,8 +23,6 @@ if "binary_handler" in locals():
         reload(export_mcfg)
     if "export_p3d" in locals():
         reload(export_p3d)
-    if "export_rtm" in locals():
-        reload(export_rtm)
     if "export_tbcsv" in locals():
         reload(export_tbcsv)
     if "import_armature" in locals():
@@ -37,8 +33,6 @@ if "binary_handler" in locals():
         reload(import_mcfg)
     if "import_p3d" in locals():
         reload(import_p3d)
-    if "import_rtm" in locals():
-        reload(import_rtm)
     if "import_tbcsv" in locals():
         reload(import_tbcsv)
     if "import_paa" in locals():
@@ -53,19 +47,16 @@ from . import compression
 from . import data_asc
 from . import data_p3d
 from . import data_p3d_odol
-from . import data_rtm
 from . import data_tbcsv
 from . import data_paa
 from . import export_asc
 from . import export_mcfg
 from . import export_p3d
-from . import export_rtm
 from . import export_tbcsv
 from . import import_armature
 from . import import_asc
 from . import import_mcfg
 from . import import_p3d
-from . import import_rtm
 from . import import_tbcsv
 from . import import_paa
 from . import odol_to_mlod