+++ DZObjectBuilder/props/__init__.py
@@ -1,8 +1,6 @@
 if "object" in locals():
     from importlib import reload
-    
-    if "action" in locals():
-        reload(action)
+
     if "material" in locals():
         reload(material)
     if "object" in locals():
@@ -11,7 +9,6 @@ if "object" in locals():
         reload(scene)
 
 
-from . import action
 from . import material
 from . import object
 from . import scene