+++ DZObjectBuilder/ui/import_export_armature.py
@@ -44,7 +44,7 @@ class A3OB_OP_import_armature(bpy.types.Operator, bpy_extras.io_utils.ImportHelp
         
         skeleton = scene_props.skeletons[self.skeleton_index]
         if riggingutils.bone_order_from_skeleton(skeleton) is None:
-            utils.op_report(self, {'ERROR'}, "Invalid skeleton definiton, run skeleton validation for RTM for more info")
+            utils.op_report(self, {'ERROR'}, "Invalid skeleton definiton, run skeleton validation for more info")
             return {'FINISHED'}
 
         arm.import_armature(self, skeleton)